--- sql/core/src/test/scala/org/apache/spark/sql/execution/command/AlterTableAddConstraintParseSuite.scala
@@ -1,377 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.spark.sql.execution.command
-
-import org.apache.spark.sql.catalyst.analysis.{UnresolvedAttribute, UnresolvedTable}
-import org.apache.spark.sql.catalyst.expressions.{CheckConstraint, ForeignKeyConstraint, GreaterThan, Literal, PrimaryKeyConstraint, UniqueConstraint}
-import org.apache.spark.sql.catalyst.parser.CatalystSqlParser.parsePlan
-import org.apache.spark.sql.catalyst.parser.ParseException
-import org.apache.spark.sql.catalyst.plans.logical.AddConstraint
-
-class AlterTableAddConstraintParseSuite extends ConstraintParseSuiteBase {
-
-  test("Add check constraint") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT c1 CHECK (d > 0)
-        |""".stripMargin
-    val parsed = parsePlan(sql)
-    val expected = AddConstraint(
-      UnresolvedTable(
-        Seq("a", "b", "c"),
-        "ALTER TABLE ... ADD CONSTRAINT"),
-      CheckConstraint(
-        child = GreaterThan(UnresolvedAttribute("d"), Literal(0)),
-        condition = "d > 0",
-        name = "c1"
-      ))
-    comparePlans(parsed, expected)
-  }
-
-  test("Add invalid check constraint name") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT c1-c3 CHECK (d > 0)
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "INVALID_IDENTIFIER", "42602", Map("ident" -> "c1-c3"))
-  }
-
-  test("Add invalid check constraint expression") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT c1 CHECK (d >)
-        |""".stripMargin
-    val msg = intercept[ParseException] {
-      parsePlan(sql)
-    }.getMessage
-    assert(msg.contains("Syntax error at or near ')'"))
-  }
-
-  test("Add check constraint with valid characteristic") {
-    validConstraintCharacteristics.foreach { case (enforcedStr, relyStr, characteristic) =>
-      val sql =
-        s"""
-           |ALTER TABLE a.b.c ADD CONSTRAINT c1 CHECK (d > 0) $enforcedStr $relyStr
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("a", "b", "c"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        CheckConstraint(
-          child = GreaterThan(UnresolvedAttribute("d"), Literal(0)),
-          condition = "d > 0",
-          name = "c1",
-          characteristic = characteristic
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-
-  test("Add check constraint with invalid characteristic") {
-    val combinations = Seq(
-      ("ENFORCED", "ENFORCED"),
-      ("ENFORCED", "NOT ENFORCED"),
-      ("NOT ENFORCED", "ENFORCED"),
-      ("NOT ENFORCED", "NOT ENFORCED"),
-      ("RELY", "RELY"),
-      ("RELY", "NORELY"),
-      ("NORELY", "RELY"),
-      ("NORELY", "NORELY")
-    )
-
-    combinations.foreach { case (characteristic1, characteristic2) =>
-      val sql =
-        s"ALTER TABLE a.b.c ADD CONSTRAINT c1 CHECK (d > 0) $characteristic1 $characteristic2"
-
-      val e = intercept[ParseException] {
-        parsePlan(sql)
-      }
-      val expectedContext = ExpectedContext(
-        fragment = s"CONSTRAINT c1 CHECK (d > 0) $characteristic1 $characteristic2",
-        start = 22,
-        stop = 50 + characteristic1.length + characteristic2.length
-      )
-      checkError(
-        exception = e,
-        condition = "INVALID_CONSTRAINT_CHARACTERISTICS",
-        parameters = Map("characteristics" -> s"$characteristic1, $characteristic2"),
-        queryContext = Array(expectedContext))
-    }
-  }
-
-  test("Add primary key constraint") {
-    Seq(("", null), ("CONSTRAINT pk1", "pk1")).foreach { case (constraintName, expectedName) =>
-      val sql =
-        s"""
-          |ALTER TABLE a.b.c ADD $constraintName PRIMARY KEY (id, name)
-          |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("a", "b", "c"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        PrimaryKeyConstraint(
-          name = expectedName,
-          columns = Seq("id", "name")
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add invalid primary key constraint name") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT pk-1 PRIMARY KEY (id)
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "'-'", "hint" -> ""))
-  }
-
-  test("Add primary key constraint with empty columns") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT pk1 PRIMARY KEY ()
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "')'", "hint" -> ""))
-  }
-
-  test("Add primary key constraint with valid characteristic") {
-    validConstraintCharacteristics.foreach { case (enforcedStr, relyStr, characteristic) =>
-      val sql =
-        s"""
-           |ALTER TABLE a.b.c ADD CONSTRAINT pk1 PRIMARY KEY (id) $enforcedStr $relyStr
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("a", "b", "c"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        PrimaryKeyConstraint(
-          name = "pk1",
-          columns = Seq("id"),
-          characteristic = characteristic
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add primary key constraint with invalid characteristic") {
-    invalidConstraintCharacteristics.foreach { case (characteristic1, characteristic2) =>
-      val sql =
-        s"ALTER TABLE a.b.c ADD CONSTRAINT pk1 PRIMARY KEY (id) $characteristic1 $characteristic2"
-
-      val e = intercept[ParseException] {
-        parsePlan(sql)
-      }
-      val expectedContext = ExpectedContext(
-        fragment = s"CONSTRAINT pk1 PRIMARY KEY (id) $characteristic1 $characteristic2",
-        start = 22,
-        stop = 54 + characteristic1.length + characteristic2.length
-      )
-      checkError(
-        exception = e,
-        condition = "INVALID_CONSTRAINT_CHARACTERISTICS",
-        parameters = Map("characteristics" -> s"$characteristic1, $characteristic2"),
-        queryContext = Array(expectedContext))
-    }
-  }
-
-  test("Add unique constraint") {
-    Seq(("", null), ("CONSTRAINT uk1", "uk1")).foreach { case (constraintName, expectedName) =>
-      val sql =
-        s"""
-           |ALTER TABLE a.b.c ADD $constraintName UNIQUE (email, username)
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("a", "b", "c"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        UniqueConstraint(
-          name = expectedName,
-          columns = Seq("email", "username")
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add invalid unique constraint name") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT uk-1 UNIQUE (email)
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "'-'", "hint" -> ""))
-  }
-
-  test("Add unique constraint with empty columns") {
-    val sql =
-      """
-        |ALTER TABLE a.b.c ADD CONSTRAINT uk1 UNIQUE ()
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "')'", "hint" -> ""))
-  }
-
-  test("Add unique constraint with valid characteristic") {
-    validConstraintCharacteristics.foreach { case (enforcedStr, relyStr, characteristic) =>
-      val sql =
-        s"""
-           |ALTER TABLE a.b.c ADD CONSTRAINT uk1 UNIQUE (email) $enforcedStr $relyStr
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("a", "b", "c"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        UniqueConstraint(
-          name = "uk1",
-          columns = Seq("email"),
-          characteristic = characteristic
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add unique constraint with invalid characteristic") {
-    invalidConstraintCharacteristics.foreach { case (characteristic1, characteristic2) =>
-      val sql =
-        s"ALTER TABLE a.b.c ADD CONSTRAINT uk1 UNIQUE (email) $characteristic1 $characteristic2"
-
-      val e = intercept[ParseException] {
-        parsePlan(sql)
-      }
-      val expectedContext = ExpectedContext(
-        fragment = s"CONSTRAINT uk1 UNIQUE (email) $characteristic1 $characteristic2",
-        start = 22,
-        stop = 52 + characteristic1.length + characteristic2.length
-      )
-      checkError(
-        exception = e,
-        condition = "INVALID_CONSTRAINT_CHARACTERISTICS",
-        parameters = Map("characteristics" -> s"$characteristic1, $characteristic2"),
-        queryContext = Array(expectedContext))
-    }
-  }
-test("Add foreign key constraint") {
-    Seq(("", null), ("CONSTRAINT fk1", "fk1")).foreach { case (constraintName, expectedName) =>
-      val sql =
-        s"""
-           |ALTER TABLE orders ADD $constraintName FOREIGN KEY (customer_id)
-           |REFERENCES customers (id)
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("orders"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        ForeignKeyConstraint(
-          name = expectedName,
-          childColumns = Seq("customer_id"),
-          parentTableId = Seq("customers"),
-          parentColumns = Seq("id")
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add invalid foreign key constraint name") {
-    val sql =
-      """
-        |ALTER TABLE orders ADD CONSTRAINT fk-1 FOREIGN KEY (customer_id)
-        |REFERENCES customers (id)
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "'-'", "hint" -> ""))
-  }
-
-  test("Add foreign key constraint with empty columns") {
-    val sql =
-      """
-        |ALTER TABLE orders ADD CONSTRAINT fk1 FOREIGN KEY ()
-        |REFERENCES customers (id)
-        |""".stripMargin
-    val e = intercept[ParseException] {
-      parsePlan(sql)
-    }
-    checkError(e, "PARSE_SYNTAX_ERROR", "42601", Map("error" -> "')'", "hint" -> ""))
-  }
-
-  test("Add foreign key constraint with valid characteristic") {
-    validConstraintCharacteristics.foreach { case (enforcedStr, relyStr, characteristic) =>
-      val sql =
-        s"""
-           |ALTER TABLE orders ADD CONSTRAINT fk1 FOREIGN KEY (customer_id)
-           |REFERENCES customers (id) $enforcedStr $relyStr
-           |""".stripMargin
-      val parsed = parsePlan(sql)
-      val expected = AddConstraint(
-        UnresolvedTable(
-          Seq("orders"),
-          "ALTER TABLE ... ADD CONSTRAINT"),
-        ForeignKeyConstraint(
-          name = "fk1",
-          childColumns = Seq("customer_id"),
-          parentTableId = Seq("customers"),
-          parentColumns = Seq("id"),
-          characteristic = characteristic
-        ))
-      comparePlans(parsed, expected)
-    }
-  }
-
-  test("Add foreign key constraint with invalid characteristic") {
-    invalidConstraintCharacteristics.foreach { case (characteristic1, characteristic2) =>
-      val sql =
-        s"""
-           |ALTER TABLE orders ADD CONSTRAINT fk1 FOREIGN KEY (customer_id)
-           |REFERENCES customers (id) $characteristic1 $characteristic2
-           |""".stripMargin
-
-      val e = intercept[ParseException] {
-        parsePlan(sql)
-      }
-      val expectedContext = ExpectedContext(
-        fragment = s"CONSTRAINT fk1 FOREIGN KEY (customer_id)\nREFERENCES customers (id) " +
-          s"$characteristic1 $characteristic2",
-        start = 24,
-        stop = 91 + characteristic1.length + characteristic2.length
-      )
-      checkError(
-        exception = e,
-        condition = "INVALID_CONSTRAINT_CHARACTERISTICS",
-        parameters = Map("characteristics" -> s"$characteristic1, $characteristic2"),
-        queryContext = Array(expectedContext))
-    }
-  }
-}
